Add files via upload
 Laser Simulation & Unit Tests
This module implements the ray-tracing engine for the Lazor puzzle solver. It simulates how laser beams travel, reflect (A blocks), absorb (B blocks), and refract (C blocks) across the board. This module realizes the gradual propagation of laser in the chessboard and the detection of boundaries. It can accurately determine the reflection direction (distinguish the top, bottom, left and right sides), and supports the logic of splitting and absorbing light. At the same time, the module has built-in unit tests, which are used to automatically verify the correctness and stability of the main functions.